Rename views with pages
